requirements.txt
aiohttp==3.6.2
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
discord.py==1.3.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
numpy==1.18.4
pylint==2.5.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
toml==0.10.1
validators==0.15.0
websockets==8.1
wrapt==1.12.1
yarl==1.4.2